api/src/config.py
from functools import lru_cache
from typing import Optional

from dotenv import load_dotenv
from pydantic import Field
from pydantic_settings import BaseSettings

load_dotenv()


class Settings(BaseSettings):
    db_url: str
    jwt_secret: str
    jwt_expire_time: int
    reset_password_token_secret: str
    verification_token_secret: str

    class Config:
        env_file = ".env"


@lru_cache()
def get_settings():
    return Settings()  # type: ignore
