api/requirements.txt
aiosqlite==0.19.0
annotated-types==0.5.0
anyio==3.7.1
asyncio==3.4.3
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
click==8.1.7
cryptography==41.0.4
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.1
fastapi-users==12.1.2
fastapi-users-db-sqlalchemy==6.0.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
makefun==1.15.1
passlib==1.7.4
pycparser==2.21
pydantic==2.3.0
pydantic-core==2.6.3
pydantic-settings==2.0.3
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
sib-api-v3-sdk==7.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
typing-extensions==4.7.1
urllib3==2.0.6
uvicorn==0.23.2
